requirements.txt
numpy
opencv-python
supervision
ultralytics
Pillow